requirements.txt
pandas
yfinance
psycopg2-binary
pytest

